image_processor/requirements.txt
watchdog
Pillow
numpy
pymongo
